# openapi.yaml
openapi: 3.0.3
info:
  title: TalentForge API Specifications
  description: Complete OpenAPI API specifications for the TalentForge performance-verified talent marketplace.
  version: 0.1.0-alpha
servers:
  - url: http://localhost:5000/api
    description: Local Dev Server
paths:
...
  /students/problems:
    get:
      summary: Get lists of problems
      parameters:
        - name: tier
          in: query
          required: false
          schema:
            type: string
        - name: domain
          in: query
          required: false
          schema:
            type: string
            enum: [cse, ece]
      responses:
        '200':
          description: Array of problems (public parameters only. hiddenTestCases omitted)
  /students/problems/{slug}:
    get:
      summary: Get specific problem detail parameters by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Problem parameters (public parameters only. hiddenTestCases omitted)
        '404':
          description: Problem not found
...
  /students/profile:
    get:
      summary: Fetch student portfolio metadata
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile records and achievements
    post:
      summary: Update student portfolio metadata
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile records updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT